+++ src/main/java/com/travelmanagement/service/ITravelerService.java
@@ -9,11 +9,11 @@ public interface ITravelerService {
 	List<TravelerResponseDTO> getAllTravelers(Integer travelerId, Integer bookingId, Integer userId, Integer packageId,
 			Integer agencyId, String bookingStatus, String keyword, int limit, int offset) throws Exception;
 
-	int getTravelerCount(Integer travelerId, Integer bookingId, Integer userId, Integer packageId, Integer agencyId,
+	long getTravelerCount(Integer travelerId, Integer bookingId, Integer userId, Integer packageId, Integer agencyId,
 			String bookingStatus, String keyword) throws Exception;
 
 
-	TravelerResponseDTO getTravelerById(int travelerId);
+	TravelerResponseDTO getTravelerById(Integer travelerId);
 
 	void updateTravelerStatus(Integer travelerId, Integer bookingId, String status) throws Exception;
